Change double to int in Main.java

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double height, double weight) {
+    public int calculate(double height, int weight) {
         double result = weight / (height * height);
         return (int) result;
     }
